% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GetAllCitations}
\alias{GetAllCitations}
\title{Get citations for all the trees currently in the cache}
\usage{
GetAllCitations()
}
\value{
A message with the list of publications
}
\description{
Get citations for all the trees currently in the cache
}

